refactor model imports
