+++ src/app/models/database/tables/meter.model.ts
@@ -1,4 +1,4 @@
-import { AbmColum } from "../../abm.model";
+import { AbmColum } from "../../components/abm.model";
 import { DbForeignKey, DbTableContext } from "../database.model";
 import { Brand, BrandDbTableContext } from "./brand.model";
 import { ActiveConstantUnit, ActiveConstantUnitDbTableContext } from "./active-constant-unit.model";
