+++ src/app/models/index.ts
@@ -12,7 +12,7 @@ export * from "./components/step-component.model";
 export * from "./components/step-state/step-state-builder.model";
 export * from "./components/step-state/step-state-execution.model";
 export * from "./components/step-state/step-state.model";
-export * from "./abm-page.model";
-export * from "./abm.model";
+export * from "./pages/abm-page.model";
+export * from "./components/abm.model";
 export * from "./database/database.model";
-export * from "./page-name.model";
+export * from "./pages/page-name.model";
